+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#2197](https://github.com/influxdb/influxdb/pull/2197): Lock server during Open().
 - [#2200](https://github.com/influxdb/influxdb/pull/2200): Re-enable Continuous Queries.
 - [#2203](https://github.com/influxdb/influxdb/pull/2203): Fix race condition on continuous queries.
+- [#2217](https://github.com/influxdb/influxdb/pull/2217): Only revert to follower if new term is greater.
 
 ## v0.9.0-rc20 [2015-04-04]